+++ paddlenlp/experimental/autonlp/auto_trainer_base.py
@@ -13,7 +13,10 @@
 # limitations under the License.
 import copy
 import datetime
+import logging
 import os
+import shutil
+import sys
 from abc import ABCMeta, abstractmethod
 from typing import Any, Callable, Dict, List, Optional, Union
 
@@ -122,11 +125,35 @@ class AutoTrainerBase(metaclass=ABCMeta):
         Performs different data checks and inferences on the training and eval datasets
         """
 
-    @abstractmethod
-    def _construct_trainable(self, train_dataset: Dataset, eval_dataset: Dataset) -> Callable:
+    def _construct_trainable(self) -> Callable:
         """
         Returns the Trainable functions that contains the main preprocessing and training logic
         """
+
+        def trainable(model_config):
+            # import is required for proper pickling
+            from paddlenlp.utils.log import logger
+
+            stdout_handler = logging.StreamHandler(sys.stdout)
+            stdout_handler.setFormatter(logger.format)
+            logger.logger.addHandler(stdout_handler)
+
+            # construct trainer
+            model_config = model_config["candidates"]
+            trainer = self._construct_trainer(model_config)
+            # train
+            trainer.train()
+            # evaluate
+            eval_metrics = trainer.evaluate()
+            # save dygraph model
+            trainer.save_model(self.save_path)
+
+            if os.path.exists(self.training_path):
+                logger.info("Removing training checkpoints to conserve disk space")
+                shutil.rmtree(self.training_path)
+            return eval_metrics
+
+        return trainable
 
     @abstractmethod
     def _compute_metrics(self, eval_preds: EvalPrediction) -> Dict[str, float]:
